geometry2: sinks
