+++ nusamai-geometry/src/compact/linestring.rs
@@ -1,30 +1,28 @@
+use crate::{Coord, Coord2d};
 use std::borrow::Cow;
-
-use super::CoordNum;
 
 /// Computer-friendly LineString
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[derive(Debug, Clone, Default, PartialEq)]
-pub struct LineString<'a, const D: usize, T: CoordNum = f64> {
+pub struct LineString<'a, T: Coord> {
     /// Coordinates of all points
     ///
     /// e.g. `[x0, y0, z0, x1, y1, z1, ...]`
     coords: Cow<'a, [T]>,
 }
 
-pub type LineString2<'a, T = f64> = LineString<'a, 2, T>;
-pub type LineString3<'a, T = f64> = LineString<'a, 3, T>;
+pub type LineString2<'a, C = f64> = LineString<'a, [C; 2]>;
+pub type LineString3<'a, C = f64> = LineString<'a, [C; 3]>;
 
-impl<'a, const D: usize, T: CoordNum> LineString<'a, D, T> {
+impl<'a, T: Coord> LineString<'a, T> {
     /// Creates an empty LineString.
     pub fn new() -> Self {
-        Default::default()
+        Self {
+            coords: Cow::Borrowed(&[]),
+        }
     }
 
     pub fn from_raw(coords: Cow<'a, [T]>) -> Self {
-        if coords.len() % D != 0 {
-            panic!("coords.len() must be a multiple of D")
-        }
         Self { coords }
     }
 
@@ -33,7 +31,7 @@ impl<'a, const D: usize, T: CoordNum> LineString<'a, D, T> {
     }
 
     /// Returns iterator over the all points in the LineString.
-    pub fn iter(&self) -> Iter<D, T> {
+    pub fn iter(&self) -> Iter<T> {
         Iter {
             slice: &self.coords,
             pos: 0,
@@ -42,7 +40,7 @@ impl<'a, const D: usize, T: CoordNum> LineString<'a, D, T> {
     }
 
     /// Returns iterator over the all points with the start point repeated.
-    pub fn iter_closed(&self) -> Iter<D, T> {
+    pub fn iter_closed(&self) -> Iter<T> {
         Iter {
             slice: &self.coords,
             pos: 0,
@@ -52,7 +50,7 @@ impl<'a, const D: usize, T: CoordNum> LineString<'a, D, T> {
 
     /// Returns the number of points in the LineString.
     pub fn len(&self) -> usize {
-        self.coords.len() / D
+        self.coords.len()
     }
 
     /// Returns `true` if the LineString is empty.
@@ -61,8 +59,8 @@ impl<'a, const D: usize, T: CoordNum> LineString<'a, D, T> {
     }
 
     /// Appends a point to the LineString.
-    pub fn push(&mut self, coord: &[T; D]) {
-        self.coords.to_mut().extend(coord);
+    pub fn push(&mut self, coord: T) {
+        self.coords.to_mut().push(coord);
     }
 
     /// Removes all points from the LineString.
@@ -71,24 +69,16 @@ impl<'a, const D: usize, T: CoordNum> LineString<'a, D, T> {
     }
 
     /// Create a new LineString by applying the given transformation to all coordinates.
-    pub fn transform<const D2: usize, T2: CoordNum>(
-        &self,
-        f: impl Fn(&[T; D]) -> [T2; D2],
-    ) -> LineString<D2, T2> {
+    pub fn transform<T2: Coord>(&self, f: impl Fn(&T) -> T2) -> LineString<T2> {
         LineString {
-            coords: self
-                .coords
-                .chunks_exact(D)
-                .flat_map(|v| f(&v.try_into().unwrap()))
-                .collect(),
+            coords: self.coords.iter().map(f).collect(),
         }
     }
 
     /// Applies the given transformation to all coordinates in the LineString.
-    pub fn transform_inplace(&mut self, mut f: impl FnMut(&[T; D]) -> [T; D]) {
-        self.coords.to_mut().chunks_exact_mut(D).for_each(|c| {
-            let transformed = f(&c.try_into().unwrap());
-            c.copy_from_slice(&transformed);
+    pub fn transform_inplace(&mut self, mut f: impl FnMut(&T) -> T) {
+        self.coords.to_mut().iter_mut().for_each(|c| {
+            *c = f(c);
         });
     }
 
@@ -97,10 +87,8 @@ impl<'a, const D: usize, T: CoordNum> LineString<'a, D, T> {
         let len = self.coords.len();
         if len > 0 {
             let data = self.coords.to_mut();
-            for i in 0..data.len() / D / 2 {
-                for j in 0..D {
-                    data.swap(i * D + j, len - (i + 1) * D + j);
-                }
+            for i in 0..data.len() / 2 {
+                data.swap(i, len - (i + 1));
             }
         }
     }
@@ -108,19 +96,17 @@ impl<'a, const D: usize, T: CoordNum> LineString<'a, D, T> {
     /// Reverses the winding order of the coordinates in the ring, preserving the first coordinate.
     pub fn reverse_ring_inplace(&mut self) {
         let len = self.coords.len();
-        if len > D {
+        if len > 1 {
             let data = self.coords.to_mut();
-            for i in 1..(data.len() / D + 1) / 2 {
-                for j in 0..D {
-                    data.swap(i * D + j, len - i * D + j);
-                }
+            for i in 1..(data.len() + 1) / 2 {
+                data.swap(i, len - i);
             }
         }
     }
 }
 
 // 2-dimensional only
-impl<'a, T: CoordNum> LineString<'a, 2, T> {
+impl<'a, T: Coord2d> LineString<'a, T> {
     /// Returns true if the ring is counter-clockwise.
     pub fn is_ccw(&self) -> bool {
         self.signed_ring_area() > 0.0
@@ -143,49 +129,51 @@ impl<'a, T: CoordNum> LineString<'a, 2, T> {
         }
         let mut area = 0.0;
         let mut ring_iter = self.iter_closed();
-        let mut prev = ring_iter.next().unwrap();
+        let mut prev = ring_iter.next().unwrap().xy();
+        // shoelace formula
         for coord in ring_iter {
-            area += (prev[0].to_f64().unwrap() * coord[1].to_f64().unwrap())
-                - (prev[1].to_f64().unwrap() * coord[0].to_f64().unwrap());
-            prev = coord;
+            let xy = coord.xy();
+            area += (prev.0 * xy.1) - (prev.1 * xy.0);
+            prev = xy;
         }
         area / 2.0
     }
 }
 
-impl<const D: usize, T: CoordNum> AsRef<[T]> for LineString<'_, D, T> {
+impl<T: Coord> AsRef<[T]> for LineString<'_, T> {
     fn as_ref(&self) -> &[T] {
         self.coords.as_ref()
     }
 }
 
-impl<'a, const D: usize, T: CoordNum> IntoIterator for &'a LineString<'_, D, T> {
-    type Item = [T; D];
-    type IntoIter = Iter<'a, D, T>;
+impl<'a, T: Coord> IntoIterator for &'a LineString<'_, T> {
+    type Item = T;
+    type IntoIter = Iter<'a, T>;
 
     fn into_iter(self) -> Self::IntoIter {
         self.iter()
     }
 }
 
-pub struct Iter<'a, const D: usize, T: CoordNum> {
+pub struct Iter<'a, T: Coord> {
     slice: &'a [T],
     pos: usize,
     close: bool,
 }
 
-impl<'a, const D: usize, T: CoordNum> Iterator for Iter<'a, D, T> {
-    type Item = [T; D];
+impl<'a, T: Coord> Iterator for Iter<'a, T> {
+    type Item = T;
 
     fn next(&mut self) -> Option<Self::Item> {
-        self.pos += D;
-        if self.pos <= self.slice.len() {
-            Some(self.slice[self.pos - D..self.pos].try_into().unwrap())
-        } else if self.close && self.slice.len() >= D && self.pos == self.slice.len() + D {
-            Some(self.slice[..D].try_into().unwrap())
+        let v = if self.pos < self.slice.len() {
+            Some(self.slice[self.pos].clone())
+        } else if self.close && !self.slice.is_empty() && self.pos == self.slice.len() {
+            Some(self.slice[0].clone())
         } else {
             None
-        }
+        };
+        self.pos += 1;
+        v
     }
 }
 
@@ -195,16 +183,20 @@ mod tests {
 
     #[test]
     fn test_line_basic() {
-        let mut line: LineString2 =
-            LineString2::from_raw((0..8).map(|v| v as f64).collect::<Vec<f64>>().into());
+        let mut line: LineString2 = LineString2::from_raw(
+            (0..4)
+                .map(|v| [v as f64, v as f64])
+                .collect::<Vec<[f64; 2]>>()
+                .into(),
+        );
         assert_eq!(line.len(), 4);
         assert!(!line.is_empty());
         for (i, coord) in line.iter().enumerate() {
             match i {
-                0 => assert_eq!(coord, [0., 1.]),
-                1 => assert_eq!(coord, [2., 3.]),
-                2 => assert_eq!(coord, [4., 5.]),
-                3 => assert_eq!(coord, [6., 7.]),
+                0 => assert_eq!(coord, [0., 0.]),
+                1 => assert_eq!(coord, [1., 1.]),
+                2 => assert_eq!(coord, [2., 2.]),
+                3 => assert_eq!(coord, [3., 3.]),
                 _ => unreachable!(),
             }
         }
@@ -213,7 +205,7 @@ mod tests {
         assert_eq!(line.len(), 0);
         assert!(line.is_empty());
 
-        line.push(&[0., 1.]);
+        line.push([0., 1.]);
         assert_eq!(line.len(), 1);
         for _c in &line {}
     }
@@ -234,16 +226,20 @@ mod tests {
 
     #[test]
     fn test_line_close() {
-        let line: LineString2 =
-            LineString2::from_raw((0..6).map(|v| v as f64).collect::<Vec<f64>>().into());
+        let line: LineString2 = LineString2::from_raw(
+            (0..3)
+                .map(|v| [v as f64, v as f64])
+                .collect::<Vec<[f64; 2]>>()
+                .into(),
+        );
         assert_eq!(line.len(), 3);
         assert!(!line.is_empty());
         for (i, coord) in line.iter_closed().enumerate() {
             match i {
-                0 => assert_eq!(coord, [0., 1.]),
-                1 => assert_eq!(coord, [2., 3.]),
-                2 => assert_eq!(coord, [4., 5.]),
-                3 => assert_eq!(coord, [0., 1.]),
+                0 => assert_eq!(coord, [0., 0.]),
+                1 => assert_eq!(coord, [1., 1.]),
+                2 => assert_eq!(coord, [2., 2.]),
+                3 => assert_eq!(coord, [0., 0.]),
                 _ => unreachable!(),
             }
         }
@@ -252,29 +248,35 @@ mod tests {
     #[test]
     fn test_transform() {
         {
-            let line = LineString2::from_raw([0., 0., 5., 0., 5., 5., 0., 5.][..].into());
+            let line = LineString2::from_raw(vec![[0., 0.], [5., 0.], [5., 5.], [0., 5.]].into());
             let new_line = line.transform(|[x, y]| [x + 2., y + 1.]);
-            assert_eq!(new_line.raw_coords(), [2., 1., 7., 1., 7., 6., 2., 6.]);
+            assert_eq!(
+                new_line.raw_coords(),
+                [[2., 1.], [7., 1.], [7., 6.], [2., 6.]]
+            );
         }
 
         {
-            let mut line = LineString2::from_raw([0., 0., 5., 0., 5., 5., 0., 5.][..].into());
+            let mut line =
+                LineString2::from_raw([[0., 0.], [5., 0.], [5., 5.], [0., 5.]][..].into());
             line.transform_inplace(|[x, y]| [x + 2., y + 1.]);
-            assert_eq!(line.raw_coords(), [2., 1., 7., 1., 7., 6., 2., 6.]);
+            assert_eq!(line.raw_coords(), [[2., 1.], [7., 1.], [7., 6.], [2., 6.]]);
         }
     }
 
     #[test]
     fn test_winding_order() {
-        let line = LineString2::from_raw(vec![0.0, 0.0, 3.0, 0.0, 3.0, 3.0, 0.0, 3.0].into());
+        let line =
+            LineString2::from_raw(vec![[0.0, 0.0], [3.0, 0.0], [3.0, 3.0], [0.0, 3.0]].into());
         assert!(line.is_ccw());
         assert!(!line.is_cw());
 
-        let line = LineString2::from_raw(vec![0.0, 0.0, 0.0, 3.0, 3.0, 3.0, 3.0, 0.0].into());
+        let line =
+            LineString2::from_raw(vec![[0.0, 0.0], [0.0, 3.0], [3.0, 3.0], [3.0, 0.0]].into());
         assert!(!line.is_ccw());
         assert!(line.is_cw());
 
-        let line = LineString2::from_raw(vec![0.0, 0.0, 0.0, 0.0].into());
+        let line = LineString2::from_raw(vec![[0.0, 0.0], [0.0, 0.0]].into());
         assert!(!line.is_ccw());
         assert!(!line.is_cw());
     }
@@ -282,40 +284,72 @@ mod tests {
     #[test]
     fn test_reverse() {
         let mut line = LineString2::from_raw(
-            vec![0.0, 0.0, 3.0, 0.0, 3.0, 1.0, 3.0, 3.0, 1.0, 3.0, 0.0, 3.0].into(),
+            vec![
+                [0.0, 0.0],
+                [3.0, 0.0],
+                [3.0, 1.0],
+                [3.0, 3.0],
+                [1.0, 3.0],
+                [0.0, 3.0],
+            ]
+            .into(),
         );
         line.reverse_inplace();
         assert_eq!(
             line.raw_coords(),
-            vec![0.0, 3.0, 1.0, 3.0, 3.0, 3.0, 3.0, 1.0, 3.0, 0.0, 0.0, 0.0]
+            vec![
+                [0.0, 3.0],
+                [1.0, 3.0],
+                [3.0, 3.0],
+                [3.0, 1.0],
+                [3.0, 0.0],
+                [0.0, 0.0]
+            ]
         );
 
-        let mut line =
-            LineString2::from_raw(vec![0.0, 0.0, 3.0, 0.0, 6.0, 0.0, 6.0, 3.0, 3.0, 3.0].into());
+        let mut line = LineString2::from_raw(
+            vec![[0.0, 0.0], [3.0, 0.0], [6.0, 0.0], [6.0, 3.0], [3.0, 3.0]].into(),
+        );
         line.reverse_inplace();
         assert_eq!(
             line.raw_coords(),
-            vec![3.0, 3.0, 6.0, 3.0, 6.0, 0.0, 3.0, 0.0, 0.0, 0.0]
+            vec![[3.0, 3.0], [6.0, 3.0], [6.0, 0.0], [3.0, 0.0], [0.0, 0.0]]
         );
     }
 
     #[test]
     fn test_ring_reverse() {
         let mut line = LineString2::from_raw(
-            vec![0.0, 0.0, 3.0, 0.0, 3.0, 1.0, 3.0, 3.0, 1.0, 3.0, 0.0, 3.0].into(),
+            vec![
+                [0.0, 0.0],
+                [3.0, 0.0],
+                [3.0, 1.0],
+                [3.0, 3.0],
+                [1.0, 3.0],
+                [0.0, 3.0],
+            ]
+            .into(),
         );
         line.reverse_ring_inplace();
         assert_eq!(
             line.raw_coords(),
-            vec![0.0, 0.0, 0.0, 3.0, 1.0, 3.0, 3.0, 3.0, 3.0, 1.0, 3.0, 0.0]
+            vec![
+                [0.0, 0.0],
+                [0.0, 3.0],
+                [1.0, 3.0],
+                [3.0, 3.0],
+                [3.0, 1.0],
+                [3.0, 0.0]
+            ]
         );
 
-        let mut line =
-            LineString2::from_raw(vec![0.0, 0.0, 3.0, 0.0, 6.0, 0.0, 6.0, 3.0, 3.0, 3.0].into());
+        let mut line = LineString2::from_raw(
+            vec![[0.0, 0.0], [3.0, 0.0], [6.0, 0.0], [6.0, 3.0], [3.0, 3.0]].into(),
+        );
         line.reverse_ring_inplace();
         assert_eq!(
             line.raw_coords(),
-            vec![0.0, 0.0, 3.0, 3.0, 6.0, 3.0, 6.0, 0.0, 3.0, 0.0]
+            vec![[0.0, 0.0], [3.0, 3.0], [6.0, 3.0], [6.0, 0.0], [3.0, 0.0]]
         );
     }
 }
