+++ nusamai-geometry/src/compact/mod.rs
@@ -10,11 +10,47 @@ pub use linestring::{LineString, LineString2, LineString3};
 pub use multi_linestring::{MultiLineString, MultiLineString2, MultiLineString3};
 pub use multi_point::{MultiPoint, MultiPoint2, MultiPoint3};
 pub use multi_polygon::{MultiPolygon, MultiPolygon2, MultiPolygon3};
-use num_traits::{Num, NumCast};
 pub use polygon::{Polygon, Polygon2, Polygon3};
+// use num_traits::{Num, NumCast};
 
-pub trait CoordNum: Num + Copy + NumCast + PartialOrd + Default + Debug {}
-impl<T: Num + Copy + NumCast + PartialOrd + Default + Debug> CoordNum for T {}
+pub trait Coord: Clone + PartialEq {}
+
+impl<const D: usize> Coord for [f64; D] {}
+impl<const D: usize> Coord for [f32; D] {}
+impl<const D: usize> Coord for [i32; D] {}
+impl<const D: usize> Coord for [i16; D] {}
+impl Coord for f64 {}
+impl Coord for f32 {}
+impl Coord for u64 {}
+impl Coord for u32 {}
+
+pub trait Coord2d: Coord {
+    fn xy(&self) -> (f64, f64);
+}
+
+impl Coord2d for [f64; 2] {
+    fn xy(&self) -> (f64, f64) {
+        (self[0], self[1])
+    }
+}
+impl Coord2d for [f32; 2] {
+    fn xy(&self) -> (f64, f64) {
+        (self[0] as f64, self[1] as f64)
+    }
+}
+impl Coord2d for [i16; 2] {
+    fn xy(&self) -> (f64, f64) {
+        (self[0] as f64, self[1] as f64)
+    }
+}
+impl Coord2d for [i32; 2] {
+    fn xy(&self) -> (f64, f64) {
+        (self[0] as f64, self[1] as f64)
+    }
+}
+
+// pub trait CoordNum: Num + Copy + NumCast + PartialOrd + Default + Debug {}
+// impl<T: Num + Copy + NumCast + PartialOrd + Default + Debug> CoordNum for T {}
 
 /// Computer-friendly Geometry
 #[cfg_attr(
@@ -23,16 +59,16 @@ impl<T: Num + Copy + NumCast + PartialOrd + Default + Debug> CoordNum for T {}
     serde(tag = "type")
 )]
 #[derive(Debug, Clone)]
-pub enum Geometry<'a, const D: usize, T: CoordNum> {
-    MultiPoint(MultiPoint<'a, D, T>),
-    LineString(LineString<'a, D, T>),
-    MultiLineString(MultiLineString<'a, D, T>),
-    Polygon(Polygon<'a, D, T>),
-    MultiPolygon(MultiPolygon<'a, D, T>),
+pub enum Geometry<'a, T: Coord> {
+    MultiPoint(MultiPoint<'a, T>),
+    LineString(LineString<'a, T>),
+    MultiLineString(MultiLineString<'a, T>),
+    Polygon(Polygon<'a, T>),
+    MultiPolygon(MultiPolygon<'a, T>),
 }
 
-pub type Geometry2<'a, T = f64> = Geometry<'a, 2, T>;
-pub type Geometry3<'a, T = f64> = Geometry<'a, 3, T>;
+pub type Geometry2<'a, C = f64> = Geometry<'a, [C; 2]>;
+pub type Geometry3<'a, C = f64> = Geometry<'a, [C; 2]>;
 
 #[cfg(test)]
 mod tests {
@@ -41,10 +77,10 @@ mod tests {
     #[test]
     fn test_coord_num_trait() {
         // 2D LineString with floating point numbers
-        let a: LineString<2, f32> = LineString::from_raw(vec![1.2, 2.3, 3.4, 4.5].into());
+        let a: LineString2<f32> = LineString::from_raw(vec![[1.2, 2.3], [3.4, 4.5]].into());
         assert_eq!(a.len(), 2);
         // Can also be used to store integer values (e.g., vertex indices)
-        let b: LineString<1, u32> = LineString::from_raw(vec![1, 2, 3].into());
+        let b: LineString<u32> = LineString::from_raw(vec![1, 2, 3].into());
         assert_eq!(b.len(), 3);
     }
 }
