--- nusamai-geometry2/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "nusamai-geometry2"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-num-traits = "0.2.18"
-serde = { version = "1.0.197", features = ["derive"], optional = true }
-
-[dev-dependencies]
-byteorder = "1.5.0"
-geo-types = "0.7.13"
-geojson = "0.24.1"
-indexmap = "2.2.5"
-quick-xml = "0.31.0"
-thiserror = "1.0.58"
-earcut-rs = { git = "https://github.com/MIERUNE/earcut-rs.git" }
-clap = { version = "4.5.3", features = ["derive"] }
-serde = { version = "1.0.197", features = ["derive"] }
-serde_json = "1.0.114"
