--- nusamai-geometry2/src/lib.rs
@@ -1,3 +1,0 @@
-mod compact;
-
-pub use compact::*;
