+++ nusamai-shapefile/src/conversion.rs
@@ -1,6 +1,6 @@
 use nusamai_geometry::{
-    CoordNum, MultiLineString, MultiLineString3, MultiPoint, MultiPoint3, MultiPolygon,
-    MultiPolygon3, Polygon, Polygon3,
+    Coord, MultiLineString, MultiLineString3, MultiPoint, MultiPoint3, MultiPolygon, MultiPolygon3,
+    Polygon, Polygon3,
 };
 use shapefile::NO_DATA;
 
@@ -12,15 +12,15 @@ pub fn polygon_to_shape(poly: &Polygon3) -> shapefile::PolygonZ {
 /// Create a Shapefile Polygon from vertices and indices.
 pub fn indexed_polygon_to_shape(
     vertices: &[[f64; 3]],
-    poly_idx: &Polygon<1, u32>,
+    poly_idx: &Polygon<u32>,
 ) -> shapefile::PolygonZ {
-    polygon_to_shape_with_mapping(poly_idx, |idx| vertices[idx[0] as usize])
+    polygon_to_shape_with_mapping(poly_idx, |idx| vertices[idx as usize])
 }
 
 /// Create a Shapefile Polygon from `nusamai_geometry::Polygon` with a mapping function.
-pub fn polygon_to_shape_with_mapping<const D: usize, T: CoordNum>(
-    poly: &Polygon<D, T>,
-    mapping: impl Fn([T; D]) -> [f64; 3],
+pub fn polygon_to_shape_with_mapping<T: Coord>(
+    poly: &Polygon<T>,
+    mapping: impl Fn(T) -> [f64; 3],
 ) -> shapefile::PolygonZ {
     let all_rings = polygon_to_shape_rings_with_mapping(poly, &mapping);
     shapefile::PolygonZ::with_rings(all_rings)
@@ -34,15 +34,15 @@ pub fn multipolygon_to_shape(mpoly: &MultiPolygon3) -> shapefile::PolygonZ {
 /// Create a Shapefile Polygon from vertices and indices.
 pub fn indexed_multipolygon_to_shape(
     vertices: &[[f64; 3]],
-    mpoly_idx: &MultiPolygon<1, u32>,
+    mpoly_idx: &MultiPolygon<u32>,
 ) -> shapefile::PolygonZ {
-    multipolygon_to_shape_with_mapping(mpoly_idx, |idx| vertices[idx[0] as usize])
+    multipolygon_to_shape_with_mapping(mpoly_idx, |idx| vertices[idx as usize])
 }
 
 /// Create a Shapefile Polygon from `nusamai_geometry::MultiPolygon` with a mapping function.
-pub fn multipolygon_to_shape_with_mapping<const D: usize, T: CoordNum>(
-    mpoly: &MultiPolygon<D, T>,
-    mapping: impl Fn([T; D]) -> [f64; 3],
+pub fn multipolygon_to_shape_with_mapping<T: Coord>(
+    mpoly: &MultiPolygon<T>,
+    mapping: impl Fn(T) -> [f64; 3],
 ) -> shapefile::PolygonZ {
     let all_rings = mpoly
         .iter()
@@ -52,9 +52,9 @@ pub fn multipolygon_to_shape_with_mapping<const D: usize, T: CoordNum>(
     shapefile::PolygonZ::with_rings(all_rings)
 }
 
-fn polygon_to_shape_rings_with_mapping<const D: usize, T: CoordNum>(
-    poly: &Polygon<D, T>,
-    mapping: impl Fn([T; D]) -> [f64; 3],
+fn polygon_to_shape_rings_with_mapping<T: Coord>(
+    poly: &Polygon<T>,
+    mapping: impl Fn(T) -> [f64; 3],
 ) -> Vec<shapefile::PolygonRing<shapefile::PointZ>> {
     let outer_points = poly
         .exterior()
@@ -88,15 +88,15 @@ pub fn multilinestring_to_shape(mls: &MultiLineString3) -> shapefile::PolylineZ 
 /// Create a Shapefile PolylineZ from vertices and indices.
 pub fn indexed_multilinestring_to_shape(
     vertices: &[[f64; 3]],
-    mls_idx: &MultiLineString<1, u32>,
+    mls_idx: &MultiLineString<u32>,
 ) -> shapefile::PolylineZ {
-    multilinestring_to_shape_with_mapping(mls_idx, |idx| vertices[idx[0] as usize])
+    multilinestring_to_shape_with_mapping(mls_idx, |idx| vertices[idx as usize])
 }
 
 /// Create a Shapefile PolylineZ from `nusamai_geometry::MultiLineString` with a mapping function.
-pub fn multilinestring_to_shape_with_mapping<const D: usize, T: CoordNum>(
-    mls: &MultiLineString<D, T>,
-    mapping: impl Fn([T; D]) -> [f64; 3],
+pub fn multilinestring_to_shape_with_mapping<T: Coord>(
+    mls: &MultiLineString<T>,
+    mapping: impl Fn(T) -> [f64; 3],
 ) -> shapefile::PolylineZ {
     let parts = mls
         .iter()
@@ -119,15 +119,15 @@ pub fn multipoint_to_shape(mpoint: &MultiPoint3) -> shapefile::MultipointZ {
 /// Create a Shapefile MultiPointZ from vertices and indices.
 pub fn indexed_multipoint_to_shape(
     vertices: &[[f64; 3]],
-    mpoint_idx: &MultiPoint<1, u32>,
+    mpoint_idx: &MultiPoint<u32>,
 ) -> shapefile::MultipointZ {
-    multipoint_to_shape_with_mapping(mpoint_idx, |idx| vertices[idx[0] as usize])
+    multipoint_to_shape_with_mapping(mpoint_idx, |idx| vertices[idx as usize])
 }
 
 /// Create a Shapefile MultiPointZ from `nusamai_geometry::MultiPoint` with a mapping function.
-pub fn multipoint_to_shape_with_mapping<const D: usize, T: CoordNum>(
-    mpoint: &MultiPoint<D, T>,
-    mapping: impl Fn([T; D]) -> [f64; 3],
+pub fn multipoint_to_shape_with_mapping<T: Coord>(
+    mpoint: &MultiPoint<T>,
+    mapping: impl Fn(T) -> [f64; 3],
 ) -> shapefile::MultipointZ {
     let shape_points = mpoint
         .iter()
@@ -203,11 +203,11 @@ mod tests {
             [3., 4., 111.],
         ];
 
-        let mut poly = Polygon::<1, u32>::new();
+        let mut poly = Polygon::<u32>::new();
         // 1st polygon
-        poly.add_ring([[0], [1], [2], [3], [0]]);
-        poly.add_ring([[4], [5], [6], [7], [4]]);
-        poly.add_ring([[8], [9], [10], [11], [8]]);
+        poly.add_ring([0, 1, 2, 3, 0]);
+        poly.add_ring([4, 5, 6, 7, 4]);
+        poly.add_ring([8, 9, 10, 11, 8]);
 
         let shape = indexed_polygon_to_shape(&vertices, &poly);
 
@@ -247,7 +247,7 @@ mod tests {
 
     #[test]
     fn test_multipolygon_to_shape() {
-        let mut mpoly = MultiPolygon::<3>::new();
+        let mut mpoly = MultiPolygon3::new();
         // 1st polygon
         mpoly.add_exterior([
             [0., 0., 0.],
@@ -346,16 +346,16 @@ mod tests {
             [4., 3., 333.],
         ];
 
-        let mut mpoly = MultiPolygon::<1, u32>::new();
+        let mut mpoly = MultiPolygon::<u32>::new();
         // 1st polygon
-        mpoly.add_exterior([[0], [1], [2], [3], [0]]);
-        mpoly.add_interior([[4], [5], [6], [7], [4]]);
-        mpoly.add_interior([[8], [9], [10], [11], [8]]);
+        mpoly.add_exterior([0, 1, 2, 3, 0]);
+        mpoly.add_interior([4, 5, 6, 7, 4]);
+        mpoly.add_interior([8, 9, 10, 11, 8]);
         // 2nd polygon
-        mpoly.add_exterior([[12], [13], [14], [15], [12]]);
-        mpoly.add_interior([[16], [17], [18], [19], [16]]);
+        mpoly.add_exterior([12, 13, 14, 15, 12]);
+        mpoly.add_interior([16, 17, 18, 19, 16]);
         // 3rd polygon
-        mpoly.add_exterior([[20], [21], [22], [23], [20]]);
+        mpoly.add_exterior([20, 21, 22, 23, 20]);
 
         let shape = indexed_multipolygon_to_shape(&vertices, &mpoly);
 
@@ -433,7 +433,7 @@ mod tests {
 
     #[test]
     fn test_multilinestring_to_shape() {
-        let mut mls = MultiLineString::<3>::new();
+        let mut mls = MultiLineString3::new();
         mls.add_linestring([[11., 12., 13.], [21., 22., 23.], [31., 32., 33.]]);
         mls.add_linestring([[111., 112., 113.], [121., 122., 123.], [131., 132., 133.]]);
 
@@ -473,10 +473,10 @@ mod tests {
             [10., 11., 333.],
         ];
 
-        let mut mls = MultiLineString::<1, u32>::new();
-        mls.add_linestring([[0], [1]]);
-        mls.add_linestring([[2], [3]]);
-        mls.add_linestring([[4], [5], [6]]);
+        let mut mls = MultiLineString::<u32>::new();
+        mls.add_linestring([0, 1]);
+        mls.add_linestring([2, 3]);
+        mls.add_linestring([4, 5, 6]);
 
         let shape = indexed_multilinestring_to_shape(&vertices, &mls);
 
@@ -507,10 +507,10 @@ mod tests {
 
     #[test]
     fn test_multipoint_to_shape() {
-        let mut mpoint = MultiPoint::<3>::new();
-        mpoint.push(&[11., 12., 13.]);
-        mpoint.push(&[21., 22., 23.]);
-        mpoint.push(&[31., 32., 33.]);
+        let mut mpoint = MultiPoint3::new();
+        mpoint.push([11., 12., 13.]);
+        mpoint.push([21., 22., 23.]);
+        mpoint.push([31., 32., 33.]);
 
         let shape = multipoint_to_shape(&mpoint);
 
@@ -532,10 +532,10 @@ mod tests {
     #[test]
     fn test_indexed_multipoint() {
         let vertices = vec![[11., 12., 13.], [21., 22., 23.], [31., 32., 33.]];
-        let mut mpoint = MultiPoint::<1, u32>::new();
-        mpoint.push(&[0]);
-        mpoint.push(&[1]);
-        mpoint.push(&[2]);
+        let mut mpoint = MultiPoint::<u32>::new();
+        mpoint.push(0);
+        mpoint.push(1);
+        mpoint.push(2);
 
         let shape = indexed_multipoint_to_shape(&vertices, &mpoint);
