+++ nusamai/src/sink/cesiumtiles/mod.rs
@@ -6,7 +6,7 @@ pub(crate) mod metadata;
 mod slice;
 mod sort;
 mod tiling;
-mod utils;
+pub(crate) mod utils;
 
 use std::{
     cmp::Ordering,
@@ -31,13 +31,13 @@ use slice::{slice_to_tiles, SlicedFeature};
 use sort::BincodeExternalChunk;
 use tiling::{TileContent, TileTree};
 
-use self::utils::calculate_normal;
 use crate::{
     get_parameter_value,
     parameters::*,
     pipeline::{Feedback, PipelineError, Receiver, Result},
     sink::{DataRequirements, DataSink, DataSinkProvider, SinkInfo},
 };
+use utils::calculate_normal;
 
 pub struct CesiumTilesSinkProvider {}
 
@@ -305,6 +305,7 @@ fn tile_writing_stage(
             };
 
             let mut earcutter = Earcut::new();
+            let mut buf3d: Vec<f64> = Vec::new();
             let mut buf2d: Vec<f64> = Vec::new(); // 2d-projected [x, y]
             let mut index_buf: Vec<u32> = Vec::new();
 
@@ -360,22 +361,26 @@ fn tile_writing_stage(
                 for (poly, orig_mat_id) in feature.polygons.iter().zip_eq(feature.polygon_material_ids.iter()) {
                     let num_outer_points = match poly.hole_indices().first() {
                         Some(&v) => v as usize,
-                        None => poly.raw_coords().len() / 5,
+                        None => poly.raw_coords().len(),
                     };
 
                     let mat = feature.materials[*orig_mat_id as usize].clone();
                     let primitive = primitives.entry(mat).or_default();
                     primitive.feature_ids.insert(feature_id as u32);
 
-                    if let Some((nx, ny, nz)) = calculate_normal(poly.exterior().raw_coords(), 5) {
-                        if project3d_to_2d(poly.raw_coords(), num_outer_points, 5, &mut buf2d) {
+                    if let Some((nx, ny, nz)) = calculate_normal(
+                        poly.exterior().iter().map(|v| [v[0], v[1], v[2]])
+                    ) {
+                        buf3d.clear();
+                        buf3d.extend(poly.raw_coords().iter().flat_map(|c| [c[0], c[1], c[2]]));
+
+                        if project3d_to_2d(&buf3d, num_outer_points, 3, &mut buf2d) {
                             // earcut
                             earcutter.earcut(&buf2d, poly.hole_indices(), 2, &mut index_buf);
 
                             // collect triangles
                             primitive.indices.extend(index_buf.iter().map(|idx| {
-                                let pos = *idx as usize * 5;
-                                let [x, y, z, u, v] = poly.raw_coords()[pos..pos + 5].try_into().unwrap();
+                                let [x, y, z, u, v] = poly.raw_coords()[*idx as usize];
                                 let vbits = [
                                     (x as f32).to_bits(),
                                     (y as f32).to_bits(),
